src/requirements.txt
bleak
keyboard
google-genai